+++ casual/casual-caller/src/test/groovy/se/laz/casual/connection/caller/LookupTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 - 2018, The casual project. All rights reserved.
+ * Copyright (c) 2017 - 2026, The casual project. All rights reserved.
  *
  * This software is licensed under the MIT license, https://opensource.org/licenses/MIT
  */
@@ -41,7 +41,7 @@ class LookupTest extends Specification
        mock.getConnectionFactory() >> {
           conFac
        }
-       mock.getJndiName() >> {
+       mock.getUniqueName() >> {
           jndiNameOne
        }
        return mock
@@ -52,7 +52,7 @@ class LookupTest extends Specification
       mock.getConnectionFactory() >> {
          conFacTwo
       }
-      mock.getJndiName() >> {
+      mock.getUniqueName() >> {
          jndiNameTwo
       }
       return mock