+++ casual/casual-caller/src/test/groovy/se/laz/casual/connection/caller/services/ServiceRouteReaderTest.groovy
@@ -2,12 +2,14 @@ package se.laz.casual.connection.caller.services
 
 import spock.lang.Specification
 
+import java.nio.file.Path
+
 class ServiceRouteReaderTest extends Specification
 {
    def'reading routes'()
    {
       when:
-      ServiceRoutes serviceRoutes = ServiceRouteReader.load('src/test/resources/service-routes.json')
+      ServiceRoutes serviceRoutes = ServiceRouteReader.load(Path.of('src/test/resources/service-routes.json').toAbsolutePath().normalize())
       then:
       !serviceRoutes.isEmpty()
       !serviceRoutes.getRoute('test-service').isEmpty()
@@ -16,7 +18,7 @@ class ServiceRouteReaderTest extends Specification
    def'non existing file'()
    {
       when:
-      ServiceRouteReader.load('src/test/resources/does-not-exist.json')
+      ServiceRouteReader.load(Path.of('src/test/resources/does-not-exist.json').toAbsolutePath().normalize())
       then:
       thrown(ServiceRouteReaderException)
    }